refactor: Renaming project files and moving them to seperate folder
