+++ README.md
@@ -93,13 +93,13 @@
 # ASSIGNMENT-5
  For the task a solution with multiple console application is created and its dependancies and build order is manged.
 
- ## Task5GreetingApp
+ ## GreetingApp
         It has the main method and it is the startup project for the solution.
         The main method prints a message to greet the user.
         It is dependent on Task5MathApp.
         The main method calls the Readinput method.
 
-## Task5MathApp
+## MathApp
         It has the Methods to get input and calls the method based on the choice entered by user for
         the math operation to be performed.
         It is dependent on Task5DisplayApp and ProjectE.
@@ -117,13 +117,13 @@
                 ChoiceofMathoperation:
                         It has the switch case which gets the option enetered by user and calls the respective method, which is in MathUtilis.
 
-## Task5DislayApp
+## DislayApp
         It has the method to recieve the result from the math operation as input and displays it as result.
 
-## Task5UtilityApp
+## UtilityApp
         It has the thankyou message to be displayed at the end of the appliaction.
 
-## ProjectE
+## ValidationApp
         It has the methods to validate the input as integer and float.
         GetInput:
                 It reads input from user and coverts it into float.
@@ -136,4 +136,4 @@
         The project Task5MathApp is dependent on the Task5DisplayApp and it calls a method in the Task5DisplayApp to display the result of math operations.
         It creates a circular dependency and can be solved by creating a new project ProjectE and having the getinput methods there and making it a dependency of Task5MathApp.
 
-![Build order](Screenshot%20(13).png)
+![Build order](BuildOrder(task5).png)
