+++ src/Assignment5/task5GreetingApp/task5.GreetingApp.csproj
@@ -15,8 +15,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\ProjectE\ProjectE.csproj" />
-    <ProjectReference Include="..\task5UtilityApp\task5UtilityApp.csproj" />
+    <ProjectReference Include="..\MathApp\task5.MathApp.csproj" />
   </ItemGroup>
 
 </Project>
